requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.13; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
async-timeout==4.0.3; python_version >= '3.7'
attrs==25.3.0; python_version >= '3.8'
beautifulsoup4==4.13.4; python_full_version >= '3.7.0'
certifi==2025.6.15; python_version >= '3.7'
charset-normalizer==3.4.2; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
duckduckgo-search==8.0.4; python_version >= '3.9'
exceptiongroup==1.3.0; python_version >= '3.7'
faiss-cpu==1.11.0; python_version >= '3.9'
frozenlist==1.7.0; python_version >= '3.9'
greenlet==3.2.3; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jiter==0.10.0; python_version >= '3.9'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
langchain==0.3.25; python_version >= '3.9'
langchain-community==0.3.25; python_version >= '3.9'
langchain-core==0.3.65; python_version >= '3.9'
langchain-openai==0.3.24; python_version >= '3.9'
langchain-tavily==0.2.3; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.8; python_version >= '3.9' and python_version < '4.0'
langgraph-cli==0.3.3; python_version >= '3.9'
langsmith==0.3.45; python_version >= '3.9'
lxml==5.4.0; python_version >= '3.6'
marshmallow==3.26.1; python_version >= '3.9'
multidict==6.5.0; python_version >= '3.9'
mypy==1.16.1; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
numpy==2.2.6; python_version >= '3.10'
openai==1.88.0; python_version >= '3.8'
orjson==3.10.18; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
primp==0.15.0; python_version >= '3.8'
propcache==0.3.2; python_version >= '3.9'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydantic-settings==2.9.1; python_version >= '3.9'
python-dotenv==1.1.0; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.4; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.7; python_version >= '3.8'
sqlalchemy==2.0.41; python_version >= '3.7'
tenacity==9.1.2; python_version >= '3.9'
tiktoken==0.9.0; python_version >= '3.9'
tomli==2.2.1; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.14.0; python_version >= '3.9'
typing-inspect==0.9.0
typing-inspection==0.4.1; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
wikipedia==1.4.0
yarl==1.20.1; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
